Update stress test to use a signed certificate (#1)
* Update stress test to use a signed certificate

Process for getting a certificate
-Create CA locally
-Upload CA cert to Pelion
-Generate device certificates
-Bootstrap devices to get production certificates
-Delete CA that was uploaded (cleanup)

To perform bootstrap the program lwm2m-bootstrapper is used. This code
is located in cmd/lwm2m-bootstrapper and was copied from:
https://github.com/pelioniot/lwm2m-bootstrapper

At the commit:
ece76dff59ca4de2ba3cda5c646820b3b13cb835

Bootstrap can take several seconds to complete. To allow for generating
a large number of certificates quickly bootstrap is done in parallel in
generate-kubeconfigs.sh.

* Retry bootstrapping if it fails

The bootstrap server sometimes fails to perform bootstrap. Attempt to
bootstrap up to 5 times before exiting.

* Delete cert with same name before upload

If a certificate with the same name exists before the CA upload then
delete it. Without this the upload would fail if there is one with the
same name.